fix: refactor asset ownership and supply mode implementations into wallet_dtos
